requirements.txt
# openai==1.35.13
# SpeechRecognition==3.10.0
# pyttsx3==2.90
# pyaudio==0.2.14

google-generativeai
google-ai-generativelanguage
SpeechRecognition
pyaudio
pyttsx3
setuptools==68.0.0
requests
playsound==1.2.2
Flask==2.3.3